+++ library/changelog.md
@@ -1,5 +1,16 @@
 # WebMan WordPress Theme Framework Changelog
 
+## 2.5.2
+
+* **Fix**: Custom styles caching issue
+
+### Files changed:
+
+	changelog.md
+	init.php
+	includes/classes/class-customize-styles.php
+
+
 ## 2.5.1
 
 * **Fix**: Undefined variable PHP error when using stylesheet generator